// src/types/fieldErrors.ts
export enum FieldErrors {
  Min = 'min',
  MoreThan = 'moreThan',
  Max = 'max',
  MinLength = 'minLength',
  MaxLength = 'maxLength',
  MinWeight = 'minWeight',
  MaxWeight = 'maxWeight',
  Required = 'required',
  ErrorDate = 'errorDate',
  ErrorEndDate = 'errorEndDate',
  OnlyEnglishChars = 'onlyEnglishCharacter',
  UpperCase = 'upperCase',
  Url = 'url',
  OnlyEthAddress = 'onlyEthAddress',
  Integer = 'integer',
  Positive = 'positive',
  ETHAddress = 'ethAddress',
  IsContractAddress = 'isContractAddress',
  wrongTokenContract = 'wrongTokenContract',
  SumOfFields = 'sumOfFields',
}